requirements.txt
python-dotenv
pika
yt_dlp
requests
black
ruff
pytest
asyncpg
aiohttp
aio_pika
aiofiles
pytest-asyncio